back-end/twitter/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-model-utils==4.3.1
djangorestframework==3.14.0
install==1.3.5
Pillow==10.0.0
postgres==4.0
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
Unipath==1.1
